requirements.txt
pandas==1.5.3
plotly==5.18.0
Pyrebase4==4.7.1
scikit_learn==1.3.0
streamlit==1.28.0
xgboost==2.0.2
matplotlib==3.7.1